refactoring, cleanup

+++ hdr/bbox.h
@@ -9,14 +9,14 @@ class bbox
 {
 public:
   bbox() { y_top=0; y_bottom=0; x_left=0; x_right=0; }
-  bbox(int yt, int yb, int xl, int xr) { y_top=yt; y_bottom=yb; x_left=xl; x_right=xr; }
+  bbox(float yt, float yb, float xl, float xr) { y_top=yt; y_bottom=yb; x_left=xl; x_right=xr; }
 
   bool includes(vertex const *v) const;
 
-  int16_t y_top;
-  int16_t y_bottom;
-  int16_t x_left;
-  int16_t x_right;
+  float y_top;
+  float y_bottom;
+  float x_left;
+  float x_right;
 };
 
 #endif
